[Fix] 문구, 폰트, 크기, 아이콘 수정 및 로그아웃 버그 수정 (#174)

+++ Projects/Features/Scene/MyPageScene/MyPage/MyPageCore.swift
@@ -209,9 +209,11 @@ public struct MyPageCore {
         
       case .logout:
         return .run(
-          operation: { send in
+          operation: { [state] send in
             try await keyChainClient.deleteUserInfo()
-            try await keyChainClient.deleteRefreshToken()
+            if case .apple = state.userInfo.loginType {
+              try await keyChainClient.deleteRefreshToken()
+            }
             await send(.backToLogin)
           },
           catch: { error, send in
